# release-20.2: sql: fix performance regression due to increased hash allocations