--- a/pkg/sql/conn_executor.go
+++ b/pkg/sql/conn_executor.go
@@ -13,7 +13,6 @@ package sql
 import (
 	"context"
 	"fmt"
-	"hash"
 	"io"
 	"math"
 	"strings"
@@ -1057,7 +1056,7 @@ type connExecutor struct {
 		// transactionStatementIDs are capped to prevent unbound expansion, but we
 		// still need the statementID hash to disambiguate beyond the capped
 		// statements.
-		transactionStatementsHash hash.Hash
+		transactionStatementsHash util.FNV64
 	}
 
 	// sessionData contains the user-configurable connection variables.