--- a/pkg/sql/crdb_internal.go
+++ b/pkg/sql/crdb_internal.go
@@ -17,6 +17,7 @@ import (
 	"net"
 	"net/url"
 	"sort"
+	"strconv"
 	"strings"
 	"time"
 
@@ -886,6 +887,9 @@ CREATE TABLE crdb_internal.node_statement_statistics (
 	},
 }
 
+// TODO(arul): Explore updating the schema below to have key be an INT and
+// statement_ids be INT[] now that we've moved to having uint64 as the type of
+// StmtID and TxnKey. Issue #55284
 var crdbInternalTransactionStatisticsTable = virtualSchemaTable{
 	comment: `finer-grained transaction statistics (in-memory, not durable; local node only). ` +
 		`This table is wiped periodically (by default, at least every two hours)`,
@@ -962,7 +966,7 @@ CREATE TABLE crdb_internal.node_transaction_statistics (
 				}
 				stmtIDsDatum := tree.NewDArray(types.String)
 				for _, stmtID := range s.statementIDs {
-					if err := stmtIDsDatum.Append(tree.NewDString(string(stmtID))); err != nil {
+					if err := stmtIDsDatum.Append(tree.NewDString(strconv.FormatUint(uint64(stmtID), 10))); err != nil {
 						return err
 					}
 				}
@@ -972,7 +976,7 @@ CREATE TABLE crdb_internal.node_transaction_statistics (
 				err := addRow(
 					tree.NewDInt(tree.DInt(nodeID)),
 					tree.NewDString(appName),
-					tree.NewDString(string(txnKey)),
+					tree.NewDString(strconv.FormatUint(uint64(txnKey), 10)),
 					stmtIDsDatum,
 					tree.NewDInt(tree.DInt(s.mu.data.Count)),
 					tree.NewDInt(tree.DInt(s.mu.data.MaxRetries)),