… docs/generated/http/full.md
… pkg/server/serverpb/status.pb.go
… pkg/server/status_test.go
… pkg/sql/logictest/testdata/logic_test/crdb_internal
… pkg/ui/yarn.lock
… pkg/util/hash.go